Updated database variable to direct to user collection with specific subcollections rather than two general collection
