+++ index.html
@@ -4,10 +4,10 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" type="text/css" href="index.css" />
-  <script src="./index.js" defer></script>
-  <script src="./swimlaneListener.js" defer></script>
-  <script src="./swimlaneDaD.js" defer></script>
-  <script src="./todoListener.js" defer></script>
+  <script src="./index.js" defer type="module"></script>
+  <!-- <script src="./swimlaneListener.js" defer></script> -->
+  <script src="./swimlaneDaD.js" defer type="module"></script>
+  <!-- <script src="./todoListener.js" defer></script> -->
   <title>To-Done</title>
 </head>
 <body>
